create Snackbar component
